refactoring dependent facet logic into a separate module
https://coveord.atlassian.net/browse/JSUI-2189
